Refactor Unit and Integration test for CodeAcademy
