+++ tests/Unit/Mock_Coursescodeacademy.php
@@ -1,0 +1,7 @@
+<?php
+ return [
+  0 => "Learn How to Code",
+  1 => "Learn HTML",
+  2 => "Learn CSS",
+  3 => "Welcome To Codecademy"
+];
